+++ packages/terra-tabs/tests/jest/Tab.test.jsx
@@ -1,7 +1,18 @@
 import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { shallowWithIntl } from 'terra-enzyme-intl';
+import { v4 as uuidv4 } from 'uuid';
 import Tab from '../../src/common-tabs/_Tab';
+
+let mockSpyUuid;
+
+beforeAll(() => {
+  mockSpyUuid = jest.spyOn(uuidv4, 'v4').mockReturnValue('00000000-0000-0000-0000-000000000000');
+});
+
+afterAll(() => {
+  mockSpyUuid.mockRestore();
+});
 
 describe('Tab', () => {
   it('should render a common tab as disabled when indicated', () => {
@@ -11,6 +22,11 @@ describe('Tab', () => {
 
   it('should render a common tab with custom props', () => {
     const wrapper = shallowWithIntl(<Tab label="Default" className="customClass" variant="framework" />).dive();
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should render a closable tab', () => {
+    const wrapper = shallowWithIntl(<Tab label="Default" className="customClass" variant="framework" isClosable />).dive();
     expect(wrapper).toMatchSnapshot();
   });