+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,14 +2,39 @@
 
 ## Unreleased
 
+* Added
+  * Added test for data grid column resizing.
+
+## 1.42.0 - (October 20, 2023)
+
+* Added
+  * Added test to cover enabling zebra striping for the `terra-table` component.
+  * Added test for tab focus for scrollable tables for the `terra-table` component.
+
 * Updated
+  * Removed columnHeaderHeight from tests using the default implementation for `terra-data-grid`.
+
+## 1.41.0 - (October 17, 2023)
+
+* Added
+  * Added documentation for the new `terra-table` component.
+  
+* Changed
+  * Updated `terra-menu` examples with focus highlight styles for terra-button on closing menu with selection of menu-item.
+
+* Updated
+  * Updates examples for `terra-data-grid` to cover scenarios for focusable elements prevented from receiving focus.
   * Updated focusable cell test for `terra-data-grid` to account for focusable elements in a hidden container.
+
+* Added
+ * Added example for vertical orientation functionality for `terra-tabs`.
 
 ## 1.40.0 - (October 11, 2023)
 
 * Added
   * Added documentation for the new `terra-file-path` component.
   * Added documentation for FlowsheetDataGrid in `terra-data-grid`.
+  * Added additional documentation and implementation guide in adherence to the newly established documentation template for the examples of Worklist Data Grid in `terra-data-grid`.
 
 ## 1.39.0 - (October 3, 2023)
 
@@ -19,6 +44,7 @@
 
 * Updated
   * Updated examples for `terra-collapsible-menu-view` to use more meaningful controls and control labels.
+  * Updated focusable cell test for `terra-data-grid` to account for focusable elements in a hidden container.
   * Updated examples and tests to add support for content scrolling within the iframe.
   * Updated examples and tests for `terra-embedded-content-consumer` to show the preferred way to set the title.