+++ packages/terra-tabs/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## 7.11.0 - (October 20, 2023)
+
+* Added
+  * Added Vertical orientation for `terra-tabs`.
+  
+* Changed
+  * Removed additional screen reader phrase and fixed delete tab hint.
+
 ## 7.10.0 - (October 11, 2023)
 
 * Changed